MAC OS SUPORT V.1

+++ src/forensic.c
@@ -164,7 +164,7 @@ int scanfile(char* filename, forensicArgs * args) {
     char* changeTime = timestampToISO(st.st_ctime);
     char* modifyTime = timestampToISO(st.st_mtime);
     char fileInfo[1000];
-    sprintf(fileInfo, "%s,%s,%ld,%s,%s,%s", filename, file_type, st.st_size, file_access, changeTime, modifyTime);
+    sprintf(fileInfo, "%s,%s,%ld,%s,%s,%s", filename, file_type, (long) st.st_size, file_access, changeTime, modifyTime);
     //write(STDOUT_FILENO, fileInfo, strlen(fileInfo));
     free(file_type);
     free(file_access);
